+++ backend/src/main/java/com/daily/daily/common/exception/FileUploadFailureException.java
@@ -1,5 +1,8 @@
 package com.daily.daily.common.exception;
 
+import com.daily.daily.common.exception.core.ErrorCode;
+import com.daily.daily.common.exception.core.WrappingCheckedException;
+
 public class FileUploadFailureException extends WrappingCheckedException {
     public FileUploadFailureException(Exception checkedException) {
         super(ErrorCode.FILE_UPLOAD_FAILURE, checkedException);
